requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-utils==0.8.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
loguru==0.7.3
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
pamqp==3.3.0
propcache==0.2.1
psutil==5.9.8
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.38
sqlalchemy-service==0.2.10
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
WTForms==3.1.2
yarl==1.18.3
pillow
